# src/api/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BadBeatMods API",
    "description": "This isn't really fully complete, but its better than absolutely nothing.\n\nThis API documentation is automatically generated and therefor may not be 100% accurate and may be missing a few fields. For example, request bodies are not fully fleshed out, and may not be accurate. Full documentation is still currently a work in progress.",
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "https://bbm.saera.gay/api"
    }
  ],
...
  "paths": {
...
    "/auth": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Get logged in user information.",
        "description": "Get user information.",
        "responses": {
          "200": {
            "description": "Returns user information."
          },
          "401": {
            "description": "Unauthorized."
          }
        },
        "security": [
          {
            "bearerAuth": [],
            "cookieAuth": []
          }
        ]
      }
    },
...
    "/auth/github/callback": {
      "get": {
        "tags": [
          "Auth"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/auth/discord": {
      "get": {
        "tags": [
          "Auth"
        ],
        "description": "",
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "bearerAuth": [],
            "cookieAuth": []
          }
        ]
      }
    },
    "/auth/discord/callback": {
      "get": {
        "tags": [
          "Auth"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "bearerAuth": [],
            "cookieAuth": []
          }
        ]
      }
    },
    "/auth/logout": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Logout.",
        "description": "Logout.",
        "responses": {
          "200": {
            "description": "Logout successful."
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "bbm_session"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  }
}